Module_5/agent/requirements.txt
langchain
langchain-core
langchain-google-genai
langgraph
langgraph-checkpoint
ddgs
python-dotenv